backend/requirements.txt
# Core
websockets>=12.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastapi>=0.100.0
uvicorn
python-multipart>=0.0.6
psutil>=5.9.0
requests>=2.31.0
plotly>=5.18.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# A2A + MCP
a2a-sdk[sqlite]>=0.3.0
a2a-sdk[http-server]>=0.3.0
httpx>=0.28.1

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
arxiv
reportlab
python-jose[cryptography]
